# backend/docs/tesfafund-openapi.yaml
...
openapi: 3.0.0
info:
  version: 0.1.0
  title: TesfaFund API
  description: API specification for the TesfaFund platform.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
servers:
  - url: http://localhost:4000
...
paths:
...
  /campaigns:
    # DONE
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => Access to public campaigns with ensitive fields hidden
    get:
      tags:
        - Campaigns
      summary: Retrieves a list of a campaigns, with optional filtering and pagination based on user's authorization
      parameters:
        - name: title
          description: Campaign title to search by.
          in: query
          schema:
            type: string
        - name: status
          description: Filter by campaign status.
          in: query
          schema:
            type: string
            enum: [PENDING_REVIEW, VERIFIED, DENIED, LIVE, PAUSED, COMPLETED]
        - name: category
          description: Filter by campaign category.
          in: query
          schema:
            type: string
        - name: minLaunchDate
          description: Minimum launch date of campaign (inclusive).
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxLaunchDate
          description: Maximum launch date of campaign (inclusive).
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minSubmissionDate
          description: Minimum submission date of campaign (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxSubmissionDate
          description: Maximum submission date of campaign (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minVerificationDate
          description: Filter by minimum campaign verification date. This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxVerificationDate
          description: Filter by maximum campaign verification date. This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minDenialDate
          description: Earliest denial date of campaign application (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxDenialDate
          description: Latest denial date of campaign application (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minEndDate
          description: Filter by minimum campaign completion date.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxEndDate
          description: Filter by maximum campaign completion date.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: isPublic
          description: Filters public campaigns.
          in: query
          schema:
            type: boolean
        - name: ownerRecipientId
          description: The ID of the recipient that owns the campaign
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          description: Page number for pagination. Must be a positive integer.
          in: query
          schema:
            type: integer
        - name: limit
          description: Maximum number of items per page. Must be a positive integer.
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedList"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No campaigns found matching the criteria (also used instead of 403 forbidden).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # DONE
    # (Role: Supervisor, Owner: No) => No access
    # (Role: Recipient, Owner: Yes | No) => Full access
    # (Role: Anon, Owner: No) => No access
    post:
      tags:
        - Campaigns
      summary: Creates a campaign application for the authenticated Recipient.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/Campaign"
                - type: object
                  properties:
                    campaignDocuments:
                      type: array
                      items:
                        type: string
                        format: binary
                      description: Array of campaign document file names to upload.
                  required:
                    - campaignDocuments
      responses:
        "201":
          description: Campaign application submitted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method.
  /campaigns/{id}:
    # DONE
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => Access to public campaigns with ensitive fields hidden
    get:
      tags:
        - Campaigns
      summary: Retrieves a campaign by ID, based on user's authorization
      description: |
        Retrieves a campaign by ID, respecting requester authorization: public campaign (unauthenticated), Recipient's campaign (authenticated), or all campaigns (Supervisor).
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the campaign.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Campaign with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # DONE
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => No access
    # (Role: Anon | Recipient, Owner: No) => No access
    put:
      tags:
        - Campaigns
      summary: Updates the campaign data. Available only for supervisors.
      description: This is used exclusively by supervisors to update a campaign (without recipient involvement), or when they approve the campaign application.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the campaign to update.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Campaign"
      responses:
        "204":
          description: Successful update - No content to return.
        "400":
          description: Bad request - Invalid request body or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
...
components:
  schemas:
...
    PaginatedList:
      type: object
      properties:
        page-number:
          type: integer
        page-count:
          type: integer
      required:
        - page-number
        - page-count
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        fundraisingGoal:
          type: string
        status:
          $ref: "#/components/schemas/CampaignStatus"
        category:
          type: string
        paymentInfo:
          $ref: "#/components/schemas/PaymentInfo"
        submissionDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        verificationDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        denialDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        launchDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        endDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        documentUrls:
          type: array
          items:
            type: string
        redactedDocumentUrls:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - fundraisingGoal
        - paymentInfo
        - category
        - endDate
    CampaignStatus:
      type: string
      enum:
        - PENDING_REVIEW
        - VERIFIED
        - DENIED
        - LIVE
        - PAUSED
        - COMPLETED
...
    PaymentInfo:
      type: object
      properties:
        paymentMethod:
          type: string
          description: The payment method used.
        phoneNo:
          type: string
          description: The phone number associated with the payment.
        bankAccountNo:
          type: string
          description: The bank account number for bank transfers.
        bankName:
          type: string
          description: The name of the bank.
      required:
        - paymentMethod
      anyOf:
        - required:
            - phoneNo
        - required:
            - bankAccountNo
            - bankName
...
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
      additionalProperties: true
  securitySchemes:
    bearerHttpAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token using a JWT